fix(diary): allow Diary AI feedback DTOs through gateway type registry
Diary AI feedback was failing silently in the webapp because
DiaryFeedbackResponse and DiaryPromptResponse weren't in
AiResponseTypeRegistry.AllowedTypes. The /api/v1/ai/chat endpoint
strips the assembly qualifier from the client-supplied ResponseType
and looks it up in the registry; when the lookup misses, the API
falls back to a plain-text response (no structured-output schema),
the AiGatewayClient then JSON-deserializes that prose into
DiaryFeedbackResponse, JsonException is thrown, the gateway returns
default(T), and DiaryViewer shows the 'feedback failed' toast.

- Move DiaryPromptResponse and DiaryFeedbackResponse from
  SentenceStudio.AppLib (namespace SentenceStudio.Services) to
  SentenceStudio.Shared.Models so the API project can reference
  them via typeof in the registry, matching the pattern used by
  the other AI response DTOs.
- Add both types to AiResponseTypeRegistry.AllowedTypes.
